+++ resources/views/index.blade.php
@@ -187,6 +187,20 @@
               Kumpulan <strong>Pilem</strong>, tipi show.
             </h1>
 
+            <!-- ngetrigger lazy loading -->
+            <div class="w-60 h-24 border-2 rounded-md mx-auto mt-20" hidden>
+              <div class="flex animate-pulse flex-row items-center h-full justify-center space-x-5">
+                <div class="w-12 bg-gray-300 h-12 rounded-full ">
+                </div>
+                    <div class="flex flex-col space-y-3">
+                    <div class="w-36 bg-gray-300 h-6 rounded-md ">
+                    </div>
+                    <div class="w-24 bg-gray-300 h-6 rounded-md ">
+                    </div>
+                </div>
+              </div>
+            </div>
+
             <div class="meta-wrapper">
 
               <!-- <div class="badge-wrapper">
@@ -537,62 +551,7 @@
 
           <h2 class="h2 section-title">Trending TV Series</h2>
 
-          <ul class="movies-list">
-
-
-            @foreach($tvShows['results'] as $tvShow)
-
-            <li>
-              <div class="movie-card">
-
-                <a href="{{ route('tv-details', ['id' => $tvShow['id']]) }}">
-                  <figure class="card-banner">
-                    <img src="{{ 'https://image.tmdb.org/t/p/w500' . $tvShow['poster_path'] }}" alt="tvshow">
-                  </figure>
-                </a>
-
-                <div class="title-wrapper">
-                  @if(Auth::check())
-                    <livewire:bookmarks :id_item="$tvShow['id']" :item_type="'tv'">
-                  @endif
-                  <a href="{{ route('tv-details', ['id' => $tvShow['id']]) }}">
-                    <h3 class="card-title">{{$tvShow['name']}}</h3>
-                  </a>
-
-                  <time datetime="">{{ date('Y',strtotime($tvShow['first_air_date'])) }}</time>
-                </div>
-
-                <div class="card-meta">
-                  <div class="badge badge-outline">
-                    <?php
-                      $inject_tv_byId = getTvById($tvShow['id']);
-                      echo $inject_tv_byId['status'];
-                      ?>
-                  </div>
-
-                  <div class="duration">
-                    <ion-icon name="time-outline"></ion-icon>
-
-                    <time datetime="PT47M">
-                    <?php
-                      foreach($inject_tv_byId['episode_run_time'] as $time) {
-                        echo $time . ",";
-                      }
-                    ?>  
-                    min</time>
-                  </div>
-
-                  <div class="rating">
-                    <ion-icon name="star"></ion-icon>
-
-                    <data>{{ $tvShow['vote_average'] }}</data>
-                  </div>
-                </div>
-
-              </div>
-            </li>
-
-            @endforeach
+          <livewire:trending-tvs lazy="on-load"/>
 
             <!-- <li>
               <div class="movie-card">
@@ -702,7 +661,6 @@
               </div>
             </li> -->
 
-          </ul>
 
         </div>
       </section>
